+++ packages/web3-eth-contract/src/encoding.ts
@@ -167,7 +167,10 @@ export const decodeEventABI = (
 		...result,
 		returnValues: decodeLog([...(modifiedEvent.inputs ?? [])], data.data, argTopics),
 		event: modifiedEvent.name,
-		signature: modifiedEvent.anonymous || !data.topics[0] ? undefined : data.topics[0],
+		signature:
+			modifiedEvent.anonymous || !data.topics || data.topics.length === 0 || !data.topics[0]
+				? undefined
+				: data.topics[0],
 
 		raw: {
 			data: data.data,