+++ packages/web3-eth/test/integration/defaults.test.ts
@@ -37,6 +37,7 @@ import {
 	createTempAccount,
 	getSystemTestProvider,
 	isIpc,
+	isWs,
 	itIf,
 } from '../fixtures/system_test_utils';
 
@@ -56,7 +57,6 @@ const MAX_32_SIGNED_INTEGER = 2147483647;
 describe('defaults', () => {
 	let web3Eth: Web3Eth;
 	let eth2: Web3Eth;
-	let tempEth: Web3Eth;
 	let clientUrl: string;
 	let contract: Contract<typeof BasicAbi>;
 	let deployOptions: Record<string, unknown>;
@@ -472,9 +472,35 @@ describe('defaults', () => {
 			expect(eth2.blockHeaderTimeout).toBe(4);
 		});
 
+		it('enableExperimentalFeatures', () => {
+			// default
+			expect(web3Eth.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout).toBe(
+				false,
+			);
+
+			// after set
+			web3Eth.setConfig({
+				enableExperimentalFeatures: { useSubscriptionWhenCheckingBlockTimeout: true },
+			});
+			expect(web3Eth.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout).toBe(
+				true,
+			);
+
+			// set by create new instance
+			eth2 = new Web3Eth({
+				config: {
+					enableExperimentalFeatures: { useSubscriptionWhenCheckingBlockTimeout: true },
+				},
+			});
+			expect(eth2.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout).toBe(
+				true,
+			);
+		});
+
 		it('should fallback to polling if provider support `on` but `newBlockHeaders` does not arrive in `blockHeaderTimeout` seconds', async () => {
 			const tempAcc2 = await createTempAccount();
-			tempEth = new Web3Eth(clientUrl);
+
+			const tempEth: Web3Eth = new Web3Eth(clientUrl);
 			// Ensure the provider supports subscriptions to simulate the test scenario
 			// It will cause providers that does not support subscriptions (like http) to throw exception when subscribing.
 			// This case is tested to ensure that even if an error happen at subscription,
@@ -560,8 +586,11 @@ describe('defaults', () => {
 					to,
 					value,
 					from,
-					nonce: Number.MAX_SAFE_INTEGER,
+					// Give a high nonce so the transaction stuck forever.
+					// However, make this random to be able to run the test many times without receiving an error that indicate submitting the same transaction twice.
+					nonce: Number.MAX_SAFE_INTEGER - Math.floor(Math.random() * 100000000),
 				});
+				expect(true).toBe(false); // the test should fail if there is no exception
 			} catch (error) {
 				// Some providers would not respond to the RPC request when sending a transaction (like Ganache v7.4.0)
 				if (error instanceof TransactionSendTimeoutError) {
@@ -588,18 +617,19 @@ describe('defaults', () => {
 
 		it('should fail if transaction was not mined within `transactionBlockTimeout` blocks', async () => {
 			const eth = new Web3Eth(clientUrl);
+			const tempAcc1 = await createTempAccount();
 			const tempAcc2 = await createTempAccount();
 
 			// Make the test run faster by casing the polling to start after 2 blocks
 			eth.transactionBlockTimeout = 2;
-			// Prevent transaction from stucking for a long time if the provider (like Ganache v7.4.0)
-			//	does not respond, when raising the nonce
-			eth.transactionSendTimeout = MAX_32_SIGNED_INTEGER;
-			// Increase other timeouts
-			eth.transactionPollingTimeout = MAX_32_SIGNED_INTEGER;
 
-			const from = tempAcc2.address;
-			const to = tempAcc.address;
+			// Increase other timeouts so only `transactionBlockTimeout` would be reached
+			eth.transactionSendTimeout = MAX_32_SIGNED_INTEGER;
+			eth.transactionPollingTimeout = MAX_32_SIGNED_INTEGER;
+			eth.blockHeaderTimeout = MAX_32_SIGNED_INTEGER / 1000;
+
+			const from = tempAcc1.address;
+			const to = tempAcc2.address;
 			const value = `0x0`;
 
 			// Setting a high `nonce` when sending a transaction, to cause the RPC call to stuck at the Node
@@ -610,38 +640,27 @@ describe('defaults', () => {
 				to,
 				value,
 				from,
-				// The previous test has the nonce set to Number.MAX_SAFE_INTEGER.
-				//	So, just decrease 1 from it here to not fall into another error.
-				nonce: Number.MAX_SAFE_INTEGER - 1,
+				// Give a high nonce so the transaction stuck forever.
+				// However, make this random to be able to run the test many times without receiving an error that indicate submitting the same transaction twice.
+				nonce: Number.MAX_SAFE_INTEGER - Math.floor(Math.random() * 100000000),
 			});
 
 			// Some providers (mostly used for development) will make blocks only when there are new transactions
-			// So, send 2 transactions because in this test `transactionBlockTimeout = 2`. And do nothing if an error happens.
-			setTimeout(() => {
-				(async () => {
-					try {
-						await eth.sendTransaction({
-							from: tempAcc.address,
-							to: tempAcc2.address,
-							value,
-						});
-					} catch (error) {
-						// Nothing needed to be done.
-					}
-					try {
-						await eth.sendTransaction({
-							from: tempAcc.address,
-							to: tempAcc2.address,
-							value,
-						});
-					} catch (error) {
-						// Nothing needed to be done.
-					}
-				})() as unknown;
-			}, 100);
+			// So, send 2 transactions, one after another, because in this test `transactionBlockTimeout = 2`.
+			// eslint-disable-next-line no-void
+			void sendFewTxes({
+				web3Eth: eth,
+				from: tempAcc2.address,
+				to: tempAcc1.address,
+				times: 2,
+				value: '0x1',
+			});
 
 			try {
 				await sentTx;
+				throw new Error(
+					'The test should fail if there is no exception when sending a transaction that could not be mined within transactionBlockTimeout',
+				);
 			} catch (error) {
 				// eslint-disable-next-line jest/no-conditional-expect
 				expect(error).toBeInstanceOf(TransactionBlockTimeoutError);
@@ -649,6 +668,70 @@ describe('defaults', () => {
 				expect((error as Error).message).toMatch(/was not mined within [0-9]+ blocks/);
 			}
 		});
+
+		// The code of this test case is identical to the pervious one except for `eth.enableExperimentalFeatures = true`
+		// 	And this test case will be removed once https://github.com/web3/web3.js/issues/5521 is implemented.
+		itIf(isWs)(
+			'should fail if transaction was not mined within `transactionBlockTimeout` blocks - when subscription is used',
+			async () => {
+				const eth = new Web3Eth(clientUrl);
+
+				// using subscription to get the new blocks and fire `TransactionBlockTimeoutError` is currently supported only
+				//	with `enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout` equal true.
+				eth.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout = true;
+
+				const tempAcc1 = await createTempAccount();
+				const tempAcc2 = await createTempAccount();
+
+				// Make the test run faster by casing the polling to start after 2 blocks
+				eth.transactionBlockTimeout = 2;
+
+				// Increase other timeouts so only `transactionBlockTimeout` would be reached
+				eth.transactionSendTimeout = MAX_32_SIGNED_INTEGER;
+				eth.transactionPollingTimeout = MAX_32_SIGNED_INTEGER;
+				eth.blockHeaderTimeout = MAX_32_SIGNED_INTEGER / 1000;
+
+				const from = tempAcc1.address;
+				const to = tempAcc2.address;
+				const value = `0x0`;
+
+				// Setting a high `nonce` when sending a transaction, to cause the RPC call to stuck at the Node
+				const sentTx: Web3PromiEvent<
+					TransactionReceipt,
+					SendTransactionEvents<typeof DEFAULT_RETURN_FORMAT>
+				> = eth.sendTransaction({
+					to,
+					value,
+					from,
+					// Give a high nonce so the transaction stuck forever.
+					// However, make this random to be able to run the test many times without receiving an error that indicate submitting the same transaction twice.
+					nonce: Number.MAX_SAFE_INTEGER - Math.floor(Math.random() * 100000000),
+				});
+
+				// Some providers (mostly used for development) will make blocks only when there are new transactions
+				// So, send 2 transactions, one after another, because in this test `transactionBlockTimeout = 2`.
+				// eslint-disable-next-line no-void
+				void sendFewTxes({
+					web3Eth: eth,
+					from: tempAcc2.address,
+					to: tempAcc1.address,
+					times: 2,
+					value: '0x1',
+				});
+
+				try {
+					await sentTx;
+					throw new Error(
+						'The test should fail if there is no exception when sending a transaction that could not be mined within transactionBlockTimeout',
+					);
+				} catch (error) {
+					// eslint-disable-next-line jest/no-conditional-expect, jest/no-standalone-expect
+					expect(error).toBeInstanceOf(TransactionBlockTimeoutError);
+					// eslint-disable-next-line jest/no-conditional-expect, jest/no-standalone-expect
+					expect((error as Error).message).toMatch(/was not mined within [0-9]+ blocks/);
+				}
+			},
+		);
 
 		it('maxListenersWarningThreshold test default config', () => {
 			// default
@@ -775,7 +858,7 @@ describe('defaults', () => {
 			);
 			expect(res.common.hardfork()).toBe('istanbul');
 		});
-		it('defaultCommon', async () => {
+		it('defaultCommon', () => {
 			// default
 			expect(web3Eth.defaultCommon).toBeUndefined();
 			const baseChain: ValidChains = 'mainnet';