+++ CHANGELOG.md
@@ -739,6 +739,7 @@ should use 4.0.1-alpha.0 for testing.
 #### web3-core
 
 -   If the response error was `execution reverted`, raise `ContractExecutionError` and pass the response error to it in order to be set as `innerError` (this innerError will be decoded at web3-eth-contract if its ABI was provided according to EIP-838). (#5434)
+-   Added a new configuration variable `enableExperimentalFeatures`. (#5481)
 -   `registerPlugin` method to `Web3Context` (#5393)
 -   `Web3PluginBase` exported abstract class (#5393)
 -   `Web3EthPluginBase` exported abstract class (#5393)
@@ -788,6 +789,11 @@ should use 4.0.1-alpha.0 for testing.
 #### web3-error
 
 -   Moved `SignerError` from `web3-errors/src/errors/signature_errors.ts` to `web3-errors/src/errors/transaction_errors.ts`, and renamed it to `TransactionSigningError` (#5462)
+
+#### web3-eth
+
+-   Use subscription at `rejectIfBlockTimeout` when the provider supports subscription. Implement this as an experimental feature (if `useSubscriptionWhenCheckingBlockTimeout` at `enableExperimentalFeatures` is `true`). (#5481)
+-   At some test cases, optimize some codes. (#5481)
 
 ### Fixed